\name{plot.CIF}
\alias{plot.CIF}
\title{
Plots the cumulative incidence function and the pointwise confidence interval (if any).
}
\description{
Plots the cumulative incidence function and lower and upper pointwise confidence interval (if avaiable ) for each group and each type of failure from the \code{CIF} class.
}
\usage{
\method{plot}{CIF}(x, type=NULL,...)
}

\arguments{
  \item{x}{
The \code{CIF} object generated by \code{est.CIF}.
}
  \item{type}{
The name of the group covariates (if any).
}
\item{...}{other arguments
}
}

\keyword{plot}