\name{summary.CIF}
\alias{summary.CIF}
\title{
Summary for the cumulative incidence function and the pointwise confidence interval.
}
\description{
Summarize the cumulative incidence function and lower and upper pointwise confidence interval (if avaiable ) for each group and each type of failure from the \code{CIF} class.
}
\usage{
\method{summary}{CIF}(object, group, ...)
}

\arguments{
  \item{object}{
The \code{CIF} object generated by \code{est.CIF}.
}
\item{group}{The name of the group covariates (if any).
}
\item{...}{other arguments
}
}

\details{
For each group and each type of failure, this function returns the cumulative incidence function and lower and upper pointwise confidence interval (if avaiable ) for the top 5 and bottom 5 observations.
}


\keyword{summary}

