\name{test.data2}
\alias{test.data2}
\docType{data}
\title{
test data
}
\description{
This is a subset of data from a prostate cancer trial described in Byar and Green (1980).
}

\references{
Byar, D. P., and S. B. Green. "The choice of treatment for cancer patients based on covariate information." Bulletin du cancer 67.4 (1979): 477-490.
}
\examples{
data(test.data2)
}
\keyword{datasets}

