\name{train.data}
\alias{train.data}
\docType{data}
\title{
train data
}
\description{
This is a subset of data from a prostate cancer trial described in Byar and Green (1980).
}


\references{
Byar, D. P., and S. B. Green. "The choice of treatment for cancer patients based on covariate information." Bulletin du cancer 67.4 (1979): 477-490.
}

\keyword{datasets}
